requirements.txt
tqdm
linkedin-api
sqlalchemy
python-dotenv
psycopg2